Create and use RDUEmailField
Instead of using a basic EmailField directly from wtforms, we should
create a new subclass of the RDUStringField and use that so that we
benefit from the work done on implementing our own styled form fields.
