+++ application/templates/admin/add_user.html
@@ -23,8 +23,7 @@
                 <h2 class="govuk-heading-s govuk-!-margin-bottom-0">Add user</h2>
                 <form action="{{ url_for('admin.add_user') }}" method="POST">
                     {{ form.csrf_token }}
-                    {{ render_field(form.email) }}
-
+                    {{ form.email() }}
                     {{ form.user_type(legend_class="govuk-!-font-weight-bold", fieldset_class="inline") }}
 
                     <button type="submit" class="govuk-button" name="add">Add user</button>
